@@ -3,9 +3,10 @@ import ReactDOM from "react-dom";
 import "./assets/css/main.scss";
 import {App} from "./App";
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root'));
+
+root.render(
   <React.StrictMode>
     <App />
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
